# src/docs/swagger.yaml
components:
    schemas:
        api.ErrorResponse:
            properties:
                error:
                    type: string
            required:
                - error
            type: object
        api.LogsResponse:
            properties:
                logs:
                    items:
                        type: string
                    type: array
            type: object
        api.NameResponse:
            properties:
                name:
                    type: string
            type: object
...
        types.DependencyGraph:
            properties:
                nodes:
                    additionalProperties:
                        $ref: '#/components/schemas/types.DependencyNode'
                    type: object
            type: object
...
        types.ProcessPorts:
            properties:
                name:
                    type: string
                tcp_ports:
                    items:
                        type: integer
                    type: array
                udp_ports:
                    items:
                        type: integer
                    type: array
            type: object
...
    securitySchemes:
        ApiTokenAuth:
            description: Required when API token authentication is configured; otherwise no API key is required.
            in: header
            name: X-PC-Token-Key
            type: apiKey
...
info:
    contact:
        name: Process Compose Discord Channel
        url: https://discord.gg/S4xgmRSHdC
    description: This is a sample Process Compose server.
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    title: Process Compose API
    version: "1.0"
openapi: 3.0.3
paths:
    /graph:
        get:
            description: Returns the process dependency graph with current status
            operationId: GetDependencyGraph
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/types.DependencyGraph'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get dependency graph
            tags:
                - Graph
...
    /namespace/restart/{name}:
        post:
            description: Restarts the namespace
            operationId: RestartNamespace
            parameters:
                - description: Namespace Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.NameResponse'
                    description: Restarted Namespace Name
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Restart a namespace
            tags:
                - Namespace
    /namespace/start/{name}:
        post:
            description: Starts the namespace
            operationId: StartNamespace
            parameters:
                - description: Namespace Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.NameResponse'
                    description: Started Namespace Name
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Start a namespace
            tags:
                - Namespace
    /namespace/stop/{name}:
        post:
            description: Stops the namespace
            operationId: StopNamespace
            parameters:
                - description: Namespace Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.NameResponse'
                    description: Stopped Namespace Name
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Stop a namespace
            tags:
                - Namespace
    /namespaces:
        get:
            description: Retrieves all unique namespaces
            operationId: GetNamespaces
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    type: string
                                type: array
                    description: Namespaces List
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get all namespaces
            tags:
                - Namespace
    /process:
        post:
            description: Update process
            operationId: UpdateProcess
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/types.ProcessConfig'
                description: Process configuration to update
                required: true
                x-originalParamName: process
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/types.ProcessConfig'
                    description: Updated Process Config
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Updates process configuration
            tags:
                - Process
    /process/{name}:
        get:
            description: Retrieves the given process and its status
            operationId: GetProcess
            parameters:
                - description: Process Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/types.ProcessState'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get process state
            tags:
                - Process
    /process/info/{name}:
        get:
            description: Retrieves the given process and its config
            operationId: GetProcessInfo
            parameters:
                - description: Process Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/types.ProcessConfig'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get process config
            tags:
                - Process
    /process/logs/{name}:
        delete:
            description: Truncates the process logs
            operationId: TruncateProcessLogs
            parameters:
                - description: Process Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.NameResponse'
                    description: Truncated Process Name
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Truncate process logs
            tags:
                - Process
    /process/logs/{name}/{endOffset}/{limit}:
        get:
            description: Retrieves the process logs
            operationId: GetProcessLogs
            parameters:
                - description: Process Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
                - description: Offset from the end of the log
                  in: path
                  name: endOffset
                  required: true
                  schema:
                    type: integer
                - description: Limit of lines to get (0 will get all the lines till the end)
                  in: path
                  name: limit
                  required: true
                  schema:
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.LogsResponse'
                    description: Process Logs
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get process logs
            tags:
                - Process
...
    /process/ports/{name}:
        get:
            description: Retrieves process open ports
            operationId: GetProcessPorts
            parameters:
                - description: Process Name
                  in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/types.ProcessPorts'
                    description: Process Ports
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid
            summary: Get process ports
            tags:
                - Process
...
security:
    - ApiTokenAuth: []
    - {}
servers:
    - url: http://localhost:8080/